+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+
+## [0.1.5]- 2022-11-04
 ### Added
 - Show last access time in users list page from [akshay196](https://github.com/akshay196)
 
+### Fixed
 - Fix for IdP users unable to login from [niravparikh05](https://github.com/niravparikh05)
 
 ## [0.1.4]- 2022-10-14